Refactor : JWT 기능 추가에 따른 회원관련 로직 추가

+++ fitter/backend/src/main/java/com/mk/fitter/api/namedwod/service/WodServiceImpl.java
@@ -8,8 +8,10 @@ import org.springframework.stereotype.Service;
 
 import com.mk.fitter.api.namedwod.repository.WodRecordRepository;
 import com.mk.fitter.api.namedwod.repository.WodRepository;
+import com.mk.fitter.api.namedwod.repository.dto.WodDto;
 import com.mk.fitter.api.namedwod.repository.dto.WodRecordDto;
 import com.mk.fitter.api.user.repository.UserRepository;
+import com.mk.fitter.api.user.repository.dto.UserDto;
 
 import lombok.RequiredArgsConstructor;
 
@@ -21,9 +23,20 @@ public class WodServiceImpl implements WodService {
 	private final UserRepository userRepository;
 	private final WodRepository wodRepository;
 
-	public boolean createWodRecord(WodRecordDto wodRecordDto) throws Exception {
+	@Override
+	public List<WodRecordDto> getNamedWodList(String namedWodName) {
+		WodDto byName = wodRepository.findByName(namedWodName);
+		List<WodRecordDto> byWodId = wodRecordRepository.findByWod_Id(byName.getId());
+		return byWodId;
+	}
+
+	public boolean createWodRecord(WodRecordDto wodRecordDto, int userId) throws Exception {
+		Optional<UserDto> byId = userRepository.findById(userId);
+		if (byId.isEmpty()) {
+			throw new Exception("존재하지 않는 회원입니다.");
+		}
 		WodRecordDto byWodIdAndUserId = wodRecordRepository.findByWod_IdAndUser_Id(wodRecordDto.getWod().getId(),
-			wodRecordDto.getUser().getId());
+			byId.get().getId());
 		if (byWodIdAndUserId != null) {
 			throw new Exception("이미 존재하는 기록입니다.");
 		} else {
@@ -33,13 +46,13 @@ public class WodServiceImpl implements WodService {
 	}
 
 	@Override
-	public List<WodRecordDto> getWodRecordList() {
-		return wodRecordRepository.findAll();
+	public List<WodDto> getWodList() {
+		return wodRepository.findByNamedIs(true);
 	}
 
 	@Override
-	public WodRecordDto getWodRecord(int wodId) throws Exception {
-		Optional<WodRecordDto> byId = wodRecordRepository.findById(wodId);
+	public WodRecordDto getWodRecord(int wodRecordId) throws Exception {
+		Optional<WodRecordDto> byId = wodRecordRepository.findById(wodRecordId);
 		if (byId.isPresent()) {
 			return byId.get();
 		} else {
@@ -48,19 +61,38 @@ public class WodServiceImpl implements WodService {
 	}
 
 	@Override
-	public boolean modifyWodRecord(int wodRecordId, LocalTime time) throws Exception {
+	public boolean modifyWodRecord(int wodRecordId, LocalTime time, int userId) throws Exception {
+		Optional<UserDto> byId = userRepository.findById(userId);
+		if (byId.isEmpty()) {
+			throw new Exception("존재하지 않는 회원입니다.");
+		}
 		Optional<WodRecordDto> foundWod = wodRecordRepository.findById(wodRecordId);
-		if (foundWod.isPresent()) {
-			WodRecordDto modifyWod = foundWod.get();
-			modifyWod.setTime(time);
-			return true;
-		} else {
+
+		if (foundWod.isEmpty()) {
 			throw new Exception("와드 기록이 존재하지 않습니다.");
 		}
+		WodRecordDto modifyWod = foundWod.get();
+		if (modifyWod.getUser().getId() != userId) {
+			throw new Exception("본인이 작성한 기록이 아닙니다.");
+		}
+		modifyWod.setTime(time);
+		return true;
 	}
 
 	@Override
-	public boolean deleteWodRecord(int wodRecordId) {
+	public boolean deleteWodRecord(int wodRecordId, int userId) throws Exception {
+		Optional<UserDto> findUser = userRepository.findById(userId);
+		if (findUser.isEmpty()) {
+			throw new Exception("존재하지 않는 회원입니다.");
+		}
+		Optional<WodRecordDto> findWod = wodRecordRepository.findById(wodRecordId);
+		if (findWod.isEmpty()) {
+			throw new Exception("존재하지 않는 기록입니다.");
+		}
+		if (userId != findWod.get().getUser().getId()) {
+			throw new Exception("본인이 작성한 기록이 아닙니다.");
+		}
 		return wodRecordRepository.deleteById(wodRecordId);
 	}
+
 }
